refactor: split block parsing
